+++ src/js/colors.js
@@ -1,8 +1,10 @@
+"use strict";
+
 /* global rgbToHex */
 
 /**
  * --------------------------------------------------------------------------
- * CoreUI Free Boostrap Admin Template (v2.1.12): colors.js
+ * CoreUI Free Boostrap Admin Template (v2.1.15): colors.js
  * Licensed under MIT (https://coreui.io/license)
  * --------------------------------------------------------------------------
  */