+++ src/js/tooltips.js
@@ -1,6 +1,8 @@
+"use strict";
+
 /**
  * --------------------------------------------------------------------------
- * CoreUI Free Boostrap Admin Template (v2.1.12): tooltips.js
+ * CoreUI Free Boostrap Admin Template (v2.1.15): tooltips.js
  * Licensed under MIT (https://coreui.io/license)
  * --------------------------------------------------------------------------
  */