+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+# Backpack HTML Template v0.2.0
+
+## Added
+- Coming soon
+
+## Updated
+#### dependencies
+- `@core-ui/core-ui` to `^2.1.16`
+- `@coreui/coreui-plugin-chartjs-custom-tooltips` to `^1.3.1`
+- `bootstrap` to `^4.4.1`
+- `chart.js` to `^2.9.3`
+- `core-js` to `^3.1.4`
+- `flag-icon-css` to `^3.4.5`
+- `font-awesome` to `~4.7.0`
+- `perfect-scrollbar` to `^1.5.0`
+- `popper.js` to `^1.16.1`
+
+#### dev dependencies 
+- `@babel/cli` to `^7.8.3`
+- `@babel/core` to `^7.8.3`
+- `@babel/plugin-proposal-object-rest-spread` to `^7.8.3`
+- `@babel/preset-env` to `^7.8.3`
+- `autoprefixer` to `^9.7.4`
+- `copyfiles` to `^2.2.0`
+- `js-beautify` to `^1.10.3`
+- `node-sass` to `^4.13.1`
+- `nodemon` to `^1.19.4`
+- `stylelint-scss` to `^3.14.0`
+
+
+## Fixed
+- Closed sidebar does not leave enough padding #14
+- babelrc config
+- Card with switch
+- Card with accent (fixed without change the card border line)
+- Switches icon
+
+
 # Backpack HTML Template v0.1.0
 
 ## Added